fix(telemetry): cast u64→f64 for prometheus inc_by, fix opentelemetry tracer API
Pre-existing compile errors under --features production/metrics/telemetry.
prometheus::Counter::inc_by() expects f64; structured_logging was calling
.tracer() on an already-constructed Tracer (removed the double-wrap).

+++ src/telemetry/prometheus.rs
@@ -168,13 +168,13 @@ pub fn update_cache_metrics(hits: u64, misses: u64, evictions: u64, size: usize)
     CACHE_HIT_RATE.set(hit_rate);
     CACHE_SIZE.set(size as f64);
 
-    CACHE_OPERATIONS.with_label_values(&["hit"]).inc_by(hits);
+    CACHE_OPERATIONS.with_label_values(&["hit"]).inc_by(hits as f64);
 
-    CACHE_OPERATIONS.with_label_values(&["miss"]).inc_by(misses);
+    CACHE_OPERATIONS.with_label_values(&["miss"]).inc_by(misses as f64);
 
     CACHE_OPERATIONS
         .with_label_values(&["eviction"])
-        .inc_by(evictions);
+        .inc_by(evictions as f64);
 }
 
 #[cfg(feature = "metrics")]
